Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
webstack-django-jwt-auth==1.5.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg[binary]
drf-spectacular==0.27.2
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
pywebview==5.3
waitress==3.0.0